Small fix to Makefile in pacoxphfix branch. Now the make also works if the directory name is already in the 'correct' format.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/branches/ProbABEL-pacox/v.0.1-3@943 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ Makefile
@@ -35,14 +35,16 @@ $(COXREG): $(COXSRC) $(REGFILES)
 
 clean:
 	rm -f $(BINDIR)/* $(SRCDIR)/*~ $(SRCDIR)/*.o $(DOCDIR)/*~ *.zip *.tar.gz
-	rm -f ../$(BUILDDIRLNK)
+	if [ -h ../$(BUILDDIRLNK) ]; then rm ../$(BUILDDIRLNK); fi
 
 linux_distrib: clean
-	ln -s $(PWD) ../$(BUILDDIRLNK)
+	if [ ! -d ../$(BUILDDIRLNK) ]; then ln -s $(PWD) ../$(BUILDDIRLNK); fi
 	cd .. ; tar --exclude-vcs -czhvf ProbABEL_$(VERSION).tar.gz $(BUILDDIRLNK)
-	rm ../$(BUILDDIRLNK)
+	if [ -h ../$(BUILDDIRLNK) ]; then rm ../$(BUILDDIRLNK); fi
 
 win_distrib: all
-	cd .. ; zip -r9 ProbABEL_$(VERSION)_win.zip ProbABEL
+	if [ ! -d ../$(BUILDDIRLNK) ]; then ln -s $(PWD) ../$(BUILDDIRLNK); fi
+	cd .. ; zip -r9 ProbABEL_$(VERSION)_win.zip $(BUILDDIRLNK)
+	if [ -h ../$(BUILDDIRLNK) ]; then rm ../$(BUILDDIRLNK); fi
 
 distrib: linux_distrib clean win_distrib clean
